refactor: detail ui
