+++ app/src/main/java/com/example/mvvm2/ui/detail/DetailActivity.kt
@@ -1,7 +1,8 @@
-package com.example.mvvm2.ui.view
+package com.example.mvvm2.ui.detail
 
 import android.os.Bundle
 import android.util.Log
+import android.widget.Toast
 import androidx.appcompat.app.AppCompatActivity
 import androidx.databinding.DataBindingUtil
 import androidx.lifecycle.Observer
@@ -9,32 +10,33 @@ import androidx.lifecycle.ViewModelProvider
 import com.example.mvvm2.R
 import com.example.mvvm2.data.dto.ItemX
 import com.example.mvvm2.databinding.ActivityDetailBinding
-import com.example.mvvm2.viewmodel.DetailViewModel
 
 class DetailActivity : AppCompatActivity() {
-
-
-    private lateinit var binding : ActivityDetailBinding
+    private lateinit var binding: ActivityDetailBinding
     private lateinit var detailViewModel: DetailViewModel
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        initBinding()
-
         val detailItem = intent?.extras?.getParcelable<ItemX>("detailItem")
-        Log.d("DetailActivity", "clicked : $detailItem")
-        detailItem?.let { detailViewModel.updateItem(it) }
+
+        initBinding()
+        initViewModel()
+
+        detailItem?.let { detailViewModel.updateItem(it) } ?: run {
+            Toast.makeText(applicationContext, "올바르지 않은 접근입니다.", Toast.LENGTH_SHORT).show()
+            finish()
+        }
     }
 
     private fun initBinding() {
+        binding = DataBindingUtil.setContentView(this, R.layout.activity_detail)
+        binding.detailViewModel = detailViewModel
+        binding.lifecycleOwner = this
+    }
+
+    private fun initViewModel() {
         detailViewModel = ViewModelProvider(this, ViewModelProvider.NewInstanceFactory())
             .get(DetailViewModel::class.java)
-
-        detailViewModel.detailItem.observe(this, Observer {
-            binding = DataBindingUtil.setContentView(this, R.layout.activity_detail)
-            binding.detailViewModel = detailViewModel
-            binding.lifecycleOwner = this
-        })
     }
 }
